chore: fix `pnpm generate "new plugin"`
