+++ package.json
@@ -15,6 +15,7 @@
     "oxlint": "oxlint --deny-warnings --disable-nested-config --report-unused-disable-directives",
     "postgenerate": "pnpm install && pnpm format && pnpm build",
     "postinstall": "husky",
+    "pregenerate": "turbo run build --filter='./turbo/generators'",
     "pretest": "turbo run build --filter='!./dev/*'",
     "release": "changeset publish",
     "test": "vitest",
@@ -24,21 +25,16 @@
   "dependencies": {
     "@changesets/changelog-github": "^0.6.0",
     "@changesets/cli": "^2.30.0",
-    "@turbo/gen": "^2.8.16",
-    "@types/hosted-git-info": "^3.0.5",
-    "@types/validate-npm-package-name": "^4.0.2",
     "@typescript/native-preview": "catalog:",
     "eslint-plugin-react-hooks": "^7.0.1",
-    "hosted-git-info": "^9.0.2",
     "husky": "^9.1.7",
     "lint-staged": "^16.4.0",
     "oxfmt": "^0.41.0",
     "oxlint": "^1.56.0",
     "oxlint-tsgolint": "^0.17.0",
     "setup-npm-trusted-publish": "^1.0.3",
-    "turbo": "^2.8.16",
+    "turbo": "2.8.16",
     "typescript": "catalog:",
-    "validate-npm-package-name": "^7.0.2",
     "vitest": "^4.1.0",
     "vitest-package-exports": "^1.2.0"
   },
