+++ turbo/generators/package.json
@@ -1,3 +1,19 @@
 {
-  "type": "commonjs"
+  "name": "@repo/generators",
+  "version": "0.0.0",
+  "private": true,
+  "type": "commonjs",
+  "scripts": {
+    "build": "tsdown"
+  },
+  "devDependencies": {
+    "@sanity/tsconfig": "catalog:",
+    "@turbo/gen": "2.8.16",
+    "@types/hosted-git-info": "^3.0.5",
+    "@types/node": "catalog:",
+    "@types/validate-npm-package-name": "^4.0.2",
+    "hosted-git-info": "^9.0.2",
+    "tsdown": "^0.21.2",
+    "validate-npm-package-name": "^7.0.2"
+  }
 }
