+++ turbo/generators/tsdown.config.mts
@@ -1,0 +1,7 @@
+import {defineConfig} from 'tsdown'
+
+export default defineConfig({
+  entry: './src/config.ts',
+  dts: true,
+  deps: {neverBundle: ['@turbo/gen']},
+})
